# api/openapi.yml
openapi: 3.0.3
info:
  title: Go Template API
  version: 1.0.0
  description: Minimal API contract for the template entity service.
servers:
  - url: /api/v1
tags:
  - name: Entities
paths:
  /entities:
    post:
      tags:
        - Entities
      operationId: createEntity
      summary: Create entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntityRequest"
      responses:
        "201":
          description: Entity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
  /entities/{entityId}:
    get:
      tags:
        - Entities
      operationId: getEntity
      summary: Get entity by id
      parameters:
        - $ref: "#/components/parameters/EntityID"
      responses:
        "200":
          description: Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
components:
  parameters:
    EntityID:
      name: entityId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    400:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    401:
      description: Not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403:
      description: Not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Entity:
      type: object
      additionalProperties: false
      required:
        - id
        - field1
        - field2
      properties:
        id:
          type: integer
          format: int64
          example: 1
        field1:
          type: string
          minLength: 1
          example: value
        field2:
          type: integer
          format: int32
          example: 42
    CreateEntityRequest:
      type: object
      additionalProperties: false
      required:
        - field1
        - field2
      properties:
        field1:
          type: string
          minLength: 1
          example: value
        field2:
          type: integer
          format: int32
          example: 42
    Error:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          example: error message