# internal/server/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create entity
	// (POST /entities)
	CreateEntity(ctx echo.Context) error
	// Get entity by id
	// (GET /entities/{entityId})
	GetEntity(ctx echo.Context, entityId EntityID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// CreateEntity converts echo context to params.
func (w *ServerInterfaceWrapper) CreateEntity(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateEntity(ctx)
	return err
}

// GetEntity converts echo context to params.
func (w *ServerInterfaceWrapper) GetEntity(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "entityId" -------------
	var entityId EntityID

	err = runtime.BindStyledParameterWithOptions("simple", "entityId", ctx.Param("entityId"), &entityId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter entityId: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetEntity(ctx, entityId)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/entities", wrapper.CreateEntity)
	router.GET(baseURL+"/entities/:entityId", wrapper.GetEntity)

}

type N400JSONResponse Error

type N404JSONResponse Error

type N500JSONResponse Error

type CreateEntityRequestObject struct {
	Body *CreateEntityJSONRequestBody
}

type CreateEntityResponseObject interface {
	VisitCreateEntityResponse(w http.ResponseWriter) error
}

type CreateEntity201JSONResponse Entity

func (response CreateEntity201JSONResponse) VisitCreateEntityResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateEntity400JSONResponse struct{ N400JSONResponse }

func (response CreateEntity400JSONResponse) VisitCreateEntityResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateEntity500JSONResponse struct{ N500JSONResponse }

func (response CreateEntity500JSONResponse) VisitCreateEntityResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetEntityRequestObject struct {
	EntityId EntityID `json:"entityId"`
}

type GetEntityResponseObject interface {
	VisitGetEntityResponse(w http.ResponseWriter) error
}

type GetEntity200JSONResponse Entity

func (response GetEntity200JSONResponse) VisitGetEntityResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetEntity404JSONResponse struct{ N404JSONResponse }

func (response GetEntity404JSONResponse) VisitGetEntityResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetEntity500JSONResponse struct{ N500JSONResponse }

func (response GetEntity500JSONResponse) VisitGetEntityResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create entity
	// (POST /entities)
	CreateEntity(ctx context.Context, request CreateEntityRequestObject) (CreateEntityResponseObject, error)
	// Get entity by id
	// (GET /entities/{entityId})
	GetEntity(ctx context.Context, request GetEntityRequestObject) (GetEntityResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// CreateEntity operation middleware
func (sh *strictHandler) CreateEntity(ctx echo.Context) error {
	var request CreateEntityRequestObject

	var body CreateEntityJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CreateEntity(ctx.Request().Context(), request.(CreateEntityRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateEntity")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateEntityResponseObject); ok {
		return validResponse.VisitCreateEntityResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetEntity operation middleware
func (sh *strictHandler) GetEntity(ctx echo.Context, entityId EntityID) error {
	var request GetEntityRequestObject

	request.EntityId = entityId

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetEntity(ctx.Request().Context(), request.(GetEntityRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetEntity")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetEntityResponseObject); ok {
		return validResponse.VisitGetEntityResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}