# internal/server/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package gen

// CreateEntityRequest defines model for CreateEntityRequest.
type CreateEntityRequest struct {
	Field1 string `json:"field1"`
	Field2 int32  `json:"field2"`
}

// Entity defines model for Entity.
type Entity struct {
	Field1 string `json:"field1"`
	Field2 int32  `json:"field2"`
	Id     int64  `json:"id"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// EntityID defines model for EntityID.
type EntityID = int64

// N400 defines model for 400.
type N400 = Error

// N404 defines model for 404.
type N404 = Error

// N500 defines model for 500.
type N500 = Error

// CreateEntityJSONRequestBody defines body for CreateEntity for application/json ContentType.
type CreateEntityJSONRequestBody = CreateEntityRequest